Fixed a issue with negative values

+++ views.py
@@ -3,7 +3,7 @@ from handlers.UserManagement import search_user_by_email, validate_login, get_id
 from handlers.UserManagement import search_user_by_username, send_recovery_password, create_user
 from handlers.UserManagement import update_username, search_user_by_id, update_email, update_password
 from handlers.UserManagement import get_user_role
-from handlers.Verifiers import check_username_exists, check_email_exists
+from handlers.Verifiers import check_username_exists, check_email_exists, check_product_in_cart
 from handlers.Retrievers import get_all_products, get_product_by_id, get_product_reviews, get_cart, verify_product_id_exists
 from handlers.ProductManagement import create_product, remove_product, verify_id_exists, update_product_name, create_product_image
 from handlers.ProductManagement import update_product_description, update_product_price, update_product_category, update_product_quantity
@@ -379,28 +379,26 @@ def add_item_to_cart(product_id):
 
         user_cart = get_cart(username + "_cart")
 
-        if len(user_cart) != 0:
-            for product in user_cart:
-                if product["product_id"] == product_id:
-                    product_stock = get_product_by_id(product_id)["stock"]
-                    print("product_quantity: " + str(product["quantity"]))
-                    print("quantity: " + str(quantity))
-                    print("product_stock: " + str(product_stock))
-                    if product["quantity"] + quantity > product_stock:
-                        return jsonify({'error': 'Not enough stock.'}), 500
-                    else:
-                        # You can add code here to update the user's cart in the database
-                        set_cart_item(username + "_cart", product_id, quantity, "add")
-                        return jsonify({'message': 'Product added to the cart.'}), 200
-            return jsonify({'error': 'Product not found.'}), 500
+        for product in user_cart:
+            if product["product_id"] == product_id:
+                product_stock = get_product_by_id(product_id)["stock"]
+                print("product_quantity: " + str(product["quantity"]))
+                print("quantity: " + str(quantity))
+                print("product_stock: " + str(product_stock))
+                if product["quantity"] + quantity > product_stock:
+                    return jsonify({'error': 'Not enough stock.'}), 500
+                else:
+                    # You can add code here to update the user's cart in the database
+                    set_cart_item(username + "_cart", product_id, quantity, "add")
+                    return jsonify({'message': 'Product added to the cart.'}), 200
+                
+        product_stock = get_product_by_id(product_id)["stock"]
+        if quantity > product_stock or quantity < 0:
+            return jsonify({'error': 'Not enough stock.'}), 500
         else:
-            product_stock = get_product_by_id(product_id)["stock"]
-            if quantity > product_stock or quantity < 0:
-                return jsonify({'error': 'Not enough stock.'}), 500
-            else:
-                # You can add code here to update the user's cart in the database
-                set_cart_item(username + "_cart", product_id, quantity, "add")
-                return jsonify({'message': 'Product added to the cart.'}), 200
+            # You can add code here to update the user's cart in the database
+            set_cart_item(username + "_cart", product_id, quantity, "add")
+            return jsonify({'message': 'Product added to the cart.'}), 200
     except Exception as e:
         print(e)
         return jsonify({'error': str(e)}), 500
@@ -409,33 +407,36 @@ def add_item_to_cart(product_id):
 
 @views.route('/remove_item_cart/<int:product_id>', methods=['POST'])
 def remove_item_from_cart(product_id):
-    username = session.get("username")
-    username = username.lower()
+    username = session.get("username").lower()
+
     try:
         data = request.get_json()
         quantity = data.get('quantity')
 
-        user_cart = get_cart(username + "_cart")
+        if check_product_in_cart(username + "_cart", product_id) == False:
+            return jsonify({'error': 'Product not in cart.'}), 500
+        else:
+            user_cart = get_cart(username + "_cart")
 
-        for product in user_cart:
-            if product["quantity"] == 0:
-                print("delete")
-                # Remove the product from the cart
-                query = "DELETE FROM " + username + "_cart WHERE product_id = %s"
-                db_query(query, (product["product_id"],))
-            elif product["product_id"] == product_id:
-                product_stock = get_product_by_id(product_id)["stock"]
-                print("product_quantity: " + str(product["quantity"]))
-                print("quantity: " + str(quantity))
-                print("product_stock: " + str(product_stock))
-                if product["quantity"] - quantity < 0:
-                    return jsonify({'error': 'Not enough stock.'}), 500
-                else:
-                    # You can add code here to update the user's cart in the database
-                    set_cart_item(username + "_cart", product_id, quantity, "remove")
-                    return jsonify({'message': 'Product added to the cart.'}), 200
+            for product in user_cart:
+                if product["quantity"] == 0:
+                    print("delete")
+                    # Remove the product from the cart
+                    query = "DELETE FROM " + username + "_cart WHERE product_id = %s"
+                    db_query(query, (product["product_id"],))
+                elif product["product_id"] == product_id:
+                    product_stock = get_product_by_id(product_id)["stock"]
+                    print("product_quantity: " + str(product["quantity"]))
+                    print("quantity: " + str(quantity))
+                    print("product_stock: " + str(product_stock))
+                    if product["quantity"] - quantity < 0:
+                        return jsonify({'error': 'Not enough stock.'}), 500
+                    else:
+                        # You can add code here to update the user's cart in the database
+                        set_cart_item(username + "_cart", product_id, quantity, "remove")
+                        return jsonify({'message': 'Product added to the cart.'}), 200
 
-        return jsonify({'message': 'Product added to the cart.'}), 200
+            return jsonify({'message': 'Product added to the cart.'}), 200
     except Exception as e:
         print(e)
         return jsonify({'error': str(e)}), 500
